requirements.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
assertpy==1.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.14